fixing checks

+++ app/controllers/pages_controller.rb
@@ -97,9 +97,11 @@ class PagesController < ApplicationController
   
   def deletedonation
 	donation_id = params[:donation]
-	ActiveRecord::Base.connection.execute("delete from audit where donation_id=#{donation_id}")
-	ActiveRecord::Base.connection.execute("delete from donation where donation_id=#{donation_id}")
-	@message = 'Donation Deleted!'
+	if donation_id != ''
+		ActiveRecord::Base.connection.execute("delete from audit where donation_id=#{donation_id}")
+		ActiveRecord::Base.connection.execute("delete from donation where donation_id=#{donation_id}")
+		@message = 'Donation Deleted!'
+	end
   end
 
   def leaderlookup
